requirements.txt
Pillow
PyAudio
PyYAML
SpeechRecognition
cryptography==41.0.7
datasets
fastapi
gTTS
huggingface-hub
invoice2data
jinja2
kaggle
kivy
langdetect==1.0.9
llama-cpp-python
pandas
pdfkit
peft
playsound
prometheus-fastapi-instrumentator
pyinstaller
pytesseract
python-jose[cryptography]
requests
sentencepiece
sentry-sdk
slowapi
torch==2.0.1
transformers==4.30.0
uvicorn
whisper
